CN: Add warning for mismatched RW/W annotation
If the C type annotation on a RW/W does not match what CN expects the
pointer to already be (e.g. from a function argument), then warn the
user. It does not warn if the expression has type [char*] or [void*].

I also updated tag_defs.c to avoid this and other warnings about the
size of the char array.