Fix #53, parallelise cn-coq CI (a bit)
The cn-coq CI is slow, it takes 35 minutes to finish, which is the
slowest part of the CI by a long shot (the next slowest thing is the CN
proof, which is about 15-20 minutes).

The obvious thing to do is to parallelise it using (the wonderful)
diff-prog.py, which works, except running on the GitHub runners with 4
cores, 16GB RAM and 14GB SSD (see [here](https://docs.github.com/en/actions/using-github-hosted-runners/using-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories)),
this quickly runs out of memory and the runner gets killed. Sad.

So I tried max_workers=2 or 3 and found that 2 worked but 3 did not, and
hard-coded that for the cn-coq CI.

The other issue is that the timeout behaviour was variable on the CI so
I set a large timeout (300s), which does cost us 3 extra minutes of
execution time.

I also removed the colour from the test/run-cn-coq.sh output because
that made it annoying to copy-paste diffs.