Added a switch to 'cn test' to produce output in Tyche format. (#55)
This is a first whack at Tyche support in `cn test`. Adding the `--output-tyche` flag to the `cn test` command, *e.g.*:

```console
$ cn test --output-tyche results.jsonl ...
```

will output a file `results.jsonl` that can be loaded into Tyche. The "replica" text is used as the representation for each test case, and a hash thereof is used as the identifier. I believe the (wall clock) runtime of the tests is being measured accurately; no coverage data or test setup time is included.